refactor(analytics): simplify components and fix lint
- Remove unused primaryColor variable
- Use barrel import for getCategoryLabel
- Remove redundant borderRadius on bars inside overflow-hidden parents
- Replace netRow wrapper View with alignSelf on Text
- Remove WHAT comments on self-evident JSX elements
- Use ReadonlyArray return type on repository
- Fix import sorting (biome)
